feat: MemberController 작성

+++ src/main/java/heavysnow/heath/service/GoalService.java
@@ -3,7 +3,10 @@ package heavysnow.heath.service;
 import heavysnow.heath.domain.Goal;
 import heavysnow.heath.domain.Member;
 import heavysnow.heath.dto.GoalCreationDto;
+import heavysnow.heath.dto.GoalIdResponseDto;
 import heavysnow.heath.dto.GoalUpdateDto;
+import heavysnow.heath.exception.ForbiddenException;
+import heavysnow.heath.exception.NotFoundException;
 import heavysnow.heath.repository.GoalRepository;
 import heavysnow.heath.repository.MemberRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -22,10 +25,14 @@ public class GoalService {
     @Transactional
     // 특정 멤버에 대한 목표를 생성하는 메서드
     // 반환 타입 Long(goalId)로 변경
-    public Long createGoalForMember(Long memberId, GoalCreationDto goalDto){
+    public GoalIdResponseDto createGoalForMember(Long loginMemberId, Long memberId, GoalCreationDto goalDto){
+
         // JPA save()메서드로 GoalCreationDto에 저장된 값을 Goal 데이터 베이스에 저장
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new IllegalArgumentException("멤버를 찾을수 없습니다." + memberId));
+        Member member = memberRepository.findById(memberId).orElseThrow();
+
+        if (!loginMemberId.equals(memberId)) {
+            throw new ForbiddenException();
+        }
 
         Goal goal = Goal.builder()
                 .member(member)
@@ -34,16 +41,20 @@ public class GoalService {
                 .build();
 
         Goal savedGoal = goalRepository.save(goal);
-        return savedGoal.getId();
+        return new GoalIdResponseDto(savedGoal.getId());
     }
 
     // 수정 : 특정 멤버에 대한 목표 수정하는 메서드
     // 반환 타입 void로 변경
     @Transactional
-    public void updateGoalForMember(Long memberId, Long goalId, GoalUpdateDto updateDto){
+    public void updateGoalForMember(Long loginMemberId, Long memberId, Long goalId, GoalUpdateDto updateDto){
         // 목표 조회
-        Goal goal = goalRepository.findByIdAndMemberId(goalId, memberId)
-                .orElseThrow(() -> new EntityNotFoundException("해당 멤버에 대한 목표를 찾을 수 업습니다."));
+        Goal goal = goalRepository.findByIdAndMemberId(goalId, memberId).orElseThrow(NotFoundException::new);
+
+        if (!loginMemberId.equals(memberId)) {
+            throw new ForbiddenException();
+        }
+
 
         // 값 변경
         goal.update(updateDto.getContent(), updateDto.isAchieved());
@@ -52,10 +63,15 @@ public class GoalService {
 
     // 삭제 : 특정 멤버에 대한 목표 삭제하는 메서드
     @Transactional
-    public void deleteGoalForMember(Long memberId, Long goalId){
+    public void deleteGoalForMember(Long loginMemberId, Long memberId, Long goalId){
+
+        Member member = memberRepository.findById(memberId).orElseThrow();
         // 목표 조회
-        Goal goal = goalRepository.findByIdAndMemberId(goalId, memberId)
-                .orElseThrow(() -> new EntityNotFoundException("해당 멤버에 대한 목표를 찾을 수 업습니다."));
+        Goal goal = goalRepository.findByIdAndMemberId(goalId, memberId).orElseThrow(NotFoundException::new);
+
+        if (!loginMemberId.equals(memberId)) {
+            throw new ForbiddenException();
+        }
 
         goalRepository.delete(goal);
     }
